Add files via upload
The previous version of this program displayed multiple results and you were supposed to conclude that the number is prime 
if at least one outcome is positive. In this update, I was able to condense those results into one. Now it simply tells you if
the number is Proth or not in one statement.

The program still has the limit at 96 due to storage limits, I am still looking for a way around this, and will update it if and once I figure it out.